fix backend : correct code for checks if env is create or not

+++ src/commands/allCanisters.ts
@@ -210,32 +210,87 @@ export async function createAndInstallCanisters() {
           const dfxConfig = JSON.parse(data);
           if (canister.category == "backend") {
             let canisterName = canister.name;
-            newCanisterId = Principal?.fromText(dfxConfig[canisterName]);
-            installMode = "reInstall";
-            const envData = `VITE_CANISTER_ID_${canisterName.toUpperCase()}_API=${newCanisterId.toText()}\n`;
-            const envFilePath = path.join(process.cwd(), '.env');
-            const fileContent = fs.readFileSync(envFilePath, { encoding: 'utf-8' });
-            const variableKey = `VITE_CANISTER_ID_${canisterName.toUpperCase()}_API=`;
-            if (!fileContent.includes(variableKey)) {
-              if (fs.existsSync(envFilePath)) {
-                fs.appendFileSync(envFilePath, envData);
-              } else {
+            if (dfxConfig[canisterName]) {
+              newCanisterId = Principal?.fromText(dfxConfig[canisterName]);
+              installMode = "reInstall";
+              const envData = `VITE_CANISTER_ID_${canisterName.toUpperCase()}_API=${newCanisterId.toText()}\n`;
+              const envFilePath = path.join(process.cwd(), '.env');
+              if (!fs.existsSync(envFilePath)) {
                 fs.writeFileSync(envFilePath, envData);
+              } else {
+                const fileContent = fs.readFileSync(envFilePath, { encoding: 'utf-8' });
+                const variableKey = `VITE_CANISTER_ID_${canisterName.toUpperCase()}_API=`;
+                if (!fileContent.includes(variableKey)) {
+                  if (fs.existsSync(envFilePath)) {
+                    fs.appendFileSync(envFilePath, envData);
+                  } else {
+                    fs.writeFileSync(envFilePath, envData);
+                  }
+                }
+              }
+            } else {
+              let canisterName = canister.name;
+              const actor = await createCanisterActor();
+              let data: any = await actor?.get_canister_id();
+              newCanisterId = data.Ok;
+              const envData = `VITE_CANISTER_ID_${canisterName.toUpperCase()}_BACKEND=${newCanisterId?.toText()}\n`;
+              const variableKey = `VITE_CANISTER_ID_${canisterName.toUpperCase()}_API=`;
+
+              const envFilePath = path.join(process.cwd(), '.env');
+              if (!fs.existsSync(envFilePath)) {
+                fs.writeFileSync(envFilePath, envData);
+              } else {
+                const fileContent = fs.readFileSync(envFilePath, { encoding: 'utf-8' });
+                if (!fileContent.includes(variableKey)) {
+                  if (fs.existsSync(envFilePath)) {
+                    fs.appendFileSync(envFilePath, envData);
+                  } else {
+                    fs.writeFileSync(envFilePath, envData);
+                  }
+                }
               }
             }
           } else if (canister.category == "frontend") {
             let canisterName = canister.name;
-            newCanisterId = Principal?.fromText(dfxConfig[canisterName])
-            installMode = "reInstall";
-            const envData = `VITE_CANISTER_ID_${canisterName.toUpperCase()}_API=${newCanisterId.toText()}\n`;
-            const variableKey = `VITE_CANISTER_ID_${canisterName.toUpperCase()}_API=`;
-            const envFilePath = path.join(process.cwd(), '.env');
-            const fileContent = fs.readFileSync(envFilePath, { encoding: 'utf-8' });
-            if (!fileContent.includes(variableKey)) {
-              if (fs.existsSync(envFilePath)) {
-                fs.appendFileSync(envFilePath, envData);
-              } else {
+            if (dfxConfig[canisterName]) {
+              newCanisterId = Principal?.fromText(dfxConfig[canisterName])
+              installMode = "reInstall";
+              const envData = `VITE_CANISTER_ID_${canisterName.toUpperCase()}_API=${newCanisterId.toText()}\n`;
+              const variableKey = `VITE_CANISTER_ID_${canisterName.toUpperCase()}_API=`;
+              const envFilePath = path.join(process.cwd(), '.env');
+
+              if (!fs.existsSync(envFilePath)) {
                 fs.writeFileSync(envFilePath, envData);
+              } else {
+                const fileContent = fs.readFileSync(envFilePath, { encoding: 'utf-8' });
+                if (!fileContent.includes(variableKey)) {
+                  if (fs.existsSync(envFilePath)) {
+                    fs.appendFileSync(envFilePath, envData);
+                  } else {
+                    fs.writeFileSync(envFilePath, envData);
+                  }
+                }
+              }
+            } else {
+              let canisterName = canister.name;
+              const actor = await createCanisterActor();
+              let data: any = await actor?.get_canister_id();
+              newCanisterId = data.Ok;
+              const envData = `VITE_CANISTER_ID_${canisterName.toUpperCase()}_BACKEND=${newCanisterId?.toText()}\n`;
+              const variableKey = `VITE_CANISTER_ID_${canisterName.toUpperCase()}_API=`;
+
+              const envFilePath = path.join(process.cwd(), '.env');
+              if (!fs.existsSync(envFilePath)) {
+                fs.writeFileSync(envFilePath, envData);
+              } else {
+                const fileContent = fs.readFileSync(envFilePath, { encoding: 'utf-8' });
+                if (!fileContent.includes(variableKey)) {
+                  if (fs.existsSync(envFilePath)) {
+                    fs.appendFileSync(envFilePath, envData);
+                  } else {
+                    fs.writeFileSync(envFilePath, envData);
+                  }
+                }
               }
             }
           }
@@ -245,15 +300,20 @@ export async function createAndInstallCanisters() {
           const actor = await createCanisterActor();
           let data: any = await actor?.get_canister_id();
           newCanisterId = data.Ok;
-          const envData = `VITE_CANISTER_ID_${canisterName.toUpperCase()}_BACKEND=${newCanisterId.toText()}\n`;
+          const envData = `VITE_CANISTER_ID_${canisterName.toUpperCase()}_BACKEND=${newCanisterId?.toText()}\n`;
           const variableKey = `VITE_CANISTER_ID_${canisterName.toUpperCase()}_API=`;
+
           const envFilePath = path.join(process.cwd(), '.env');
-          const fileContent = fs.readFileSync(envFilePath, { encoding: 'utf-8' });
-          if (!fileContent.includes(variableKey)) {
-            if (fs.existsSync(envFilePath)) {
-              fs.appendFileSync(envFilePath, envData);
-            } else {
-              fs.writeFileSync(envFilePath, envData);
+          if (!fs.existsSync(envFilePath)) {
+            fs.writeFileSync(envFilePath, envData);
+          } else {
+            const fileContent = fs.readFileSync(envFilePath, { encoding: 'utf-8' });
+            if (!fileContent.includes(variableKey)) {
+              if (fs.existsSync(envFilePath)) {
+                fs.appendFileSync(envFilePath, envData);
+              } else {
+                fs.writeFileSync(envFilePath, envData);
+              }
             }
           }
         }
